+++ threads/rdft-vrank-geq1.c
@@ -18,7 +18,7 @@
  *
  */
 
-/* $Id: rdft-vrank-geq1.c,v 1.20 2006-01-05 23:23:15 athena Exp $ */
+/* $Id: rdft-vrank-geq1.c,v 1.21 2006-01-10 13:24:41 athena Exp $ */
 
 #include "threads.h"
 
@@ -218,7 +218,7 @@ void X(rdft_thr_vrank_geq1_register)(planner *p)
      /* FIXME: Should we try other vecloop_dim values? */
      static const int buddies[] = { 1, -1 };
 
-     const int nbuddies = sizeof(buddies) / sizeof(buddies[0]);
+     const int nbuddies = (int)(sizeof(buddies) / sizeof(buddies[0]));
 
      for (i = 0; i < nbuddies; ++i)
           REGISTER_SOLVER(p, mksolver(buddies[i], buddies, nbuddies));
